tdf#47065 Add new file open UI options and implement a new thread
Add new UI options when opening a locked or non-writeable document
to allow the user to be notified when such a document becomes editable
.  If the user selects "Notify", then that document is added to a list
of open documents to be checked by a thread every 60 seconds for
read/write access and whether lock file is available/obtainable. If
access is allowed for a document, then show UI dialog to the user
asking to Reload that document.  If Reload is selected by the user
then that document is reloaded with read/write access.  The checking
thread is spawned only once no matter how many "Notify" documents
there are.  The thread is spawned if not already running when a new
"Notify" document is opened, and it terminates when all "Notify"
documents have been closed or the application terminates.

Also update badstatics clang plugin to ignore new global variables
introduced.

Change-Id: I7555ce6f5df79c2c87216e0129ef3b2883c7d921
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111654
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>