tdf#140552 RTF export: fix hyperlink, in footnote, in hyperlink
Regression from commit 7d42346ba77c9c4df241ea40eaf550993ca18783
(tdf#90421 RTF export: ignore hyperlinks without an URL, 2015-04-21),
URLs can be nested in the footnote case, which requires a stack.
Otherwise the inner URL clears "the" URL and we don't close the outer
field as we believe it's empty, so it was not started.

Change-Id: I9f87ddbb7e597c413bf836eb9b58beb76722361f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111794
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>