tdf#135774 writerfilter Char highlight: revert tdf#117137
revert my 6.1 commit ac27f4e7abf5339f71d4f5f3fc09a13b25669fe4
tdf#117137: DOCX import: don't try to set grab-bag as UNO prop

The ultimate problems in bug 117137 seem to have been fixed
elsewhere and the unit test still passes after the revert.
(Tested in Word 2016 and 2003).
Also my entire tone was one of uncertainty in the review process.

In general, the grabbags should not be ignored, as they are supposed
to just contain import information that should be exported again.

In this specific case, CharGrabBag is needed to identify that
the background was imported as shading, so that it will still
be exported as shading instead of being converted into a highlight.

If this revert causes problems, then the grabbag will need
to remove the problematic items (or keep only the essential items).

Change-Id: I368fcdfca9057426aa6352f1caa7b552d87905d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111937
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>