sw: DOCX export: put fly before fieldmark start into its own run
If a fly starts at the same position as a CH_TXT_ATR_FIELDSTART, it is
anchored before the field, and written in OutFlys() before the field is
written in EndRun(), but the DOCX export reorders things in confusing
ways.

StartField_Impl() and CmdField_Impl() will actually end the current run
(after putting the field char in it) and start a new one.

So do something similar in this situation and create a new run if flys
have been processed.

Restrict this extra run to when there is actually a fly, because
otherwise a dozen tests break; this requires a new FLY_NONE result for
OutFlys() because FLY_PROCESSED is returned even if there are no flys.

Change-Id: Id469c53d07eacad3992c7c0e451ab3756e02c8fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112267
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>