VCL drop Scheduler::ProcessTaskScheduling
Just process tasks (and system events) via DoYield and actually
wake up the system timer on OSX.

This drops the testFocus unit test. There is some comment about
it in README.lifecycle, but the test's CPPUNIT_ASSERT was already
commented and mmeeks suggested to simply drop it.

Even worse: just replacing

-    Scheduler::ProcessTaskScheduling();
+    Scheduler::ProcessEventsToIdle();

results in a SIGSEGV in my local Linux build, which is a totally
unrelated crash.

Change-Id: Ie3e2a8668b8501f081706dde0ba3684801c30cc2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112761
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>