VCL force Scheduler instant timeout on task skip
If we skipped the current task processing, we need to immediately
schedule a new timeout, as we know there is a pending task. So set
nMinPeriod to 0, as it contains the period to the next pending
task, and update the system timer.

Fix regression from commit d3b498cc4732f964919fecb265085cefcc422469
("invoke idle priority timers only when actually idle").

Change-Id: Ie3fc176fade595ec7406fec0835835bd86dfbd90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112762
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>