tdf#141173 docx export: revert fdo#75604: File Corruption
It looks to me like a "this fixes my document" patch.
In the meantime, the real issue seems to have been solved,
since the unit test passes, and a round-tripped document
looks fine in MS Word 2016.

This fixes a regression in 7.2 - bug 11173.
(The regression backport to 7.1 should just be reverted.)

Change-Id: I329b3b9ca23d86acd05a0a61c719f2c3bce412f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112994
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>