lok: Send gridOffset of the shape
In core, the gridOffset is calculated based on the LogicRect's TopLeft coordinate
In online, we have the SnapRect and we calculate it based on its TopLeft coordinate
SnapRect's TopLeft and LogicRect's TopLeft match when there is no rotation
but the rotation is not applied to the LogicRect. Therefore,
what we calculate in online does not match in case of the rotated shape.
Here we can send the correct gridOffset in the selection callback.
whether the shape is rotated or not, we will always have the correct gridOffset
Note that the gridOffset is always calculated from the first selected obj

Change-Id: Icc62a94879367f9b55ad05887945393452021777
Signed-off-by: merttumer <mert.tumer@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113078
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114027
Tested-by: Jenkins