ofz#32111 Timeout
keep continuing on a short read, zeroing out the rest of the block, if
something was read, but give up if nothing could be read.

vcl/qa/cppunit/graphicfilter/data/png/pass/invalid-chunk.png remains
working after this. Though to be honest it is also unchanged if
we just png_error intead of png_warning

Change-Id: I7dc24bcfc0b33e534a3f9665b11cd9aca8003a89
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113152
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>