related tdf#140336 docxsdrexport: always clear vmlTextFrame attrs
Blind fix - it just doesn't look right, and caused troubles
when making a patch for bug 140336. When the VMLTextFrame is
written out, surely the collected attributes for it should
be cleared instead of spilling over into the next
non-textbox-only VMLTextFrame.

I expect the textbox-only frames never have any gradients or
SOLID fills, so it has never been a problem yet.

Change-Id: I3aa23aadf9ab5a0be54ffbd7ea114f94c48aaddd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113206
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>