tdf#140343 sw page rtl gutter margin: add DOC filter
The import's tokenizer was there since commit
3e13e053749e1bf19179bf83bdf0a18a9488b5e8 (#i2408# Import and export
set/ask references fields as set/input, change references to bookmarks
to show variable as appropiate, 2002-02-04), but it tried to read 2
bytes, not 1 byte, so it was broken. This did not cause problems in
practice till commit b897cc4dfc7111eb8dfd5d8aa8c970f21ab035d6
(consistently track amount of buffers remaining, 2017-04-10) started to
check for not reading past the end of the buffer.

The export was missing, since we previously tried to handle this at
import time, so the doc model didn't have the info anymore.

Change-Id: I0d15a16893de96104812b0cba4566c53f92673ae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113415
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins