vcl: move GenPspGraphics into a subclass of SalGraphicsImpl
This is an effort to make SalGraphicsImpl mandatory for all
backends. So we can rearange the interfaces of backends to a more
logical structure, where we can have multiple graphic rendering
backends for an VCL (windowing) backend (like we already do for
Windows with GDI and skia backends).

In this case we changed GenPspGraphics, introduced a new GFX
backend "GenPspGfxBackend" - a subclass of SalGraphicsImpl, and
moved the graphic functions into it.

Change-Id: I148b460ff71c3e6b7548e9e2b933390151b6d82f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113593
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>